+++ pyodide-build/pyodide_build/cli/create_zipfile.py
@@ -1,3 +1,4 @@
+import re
 from pathlib import Path
 
 import typer
@@ -9,6 +10,14 @@ def main(
     libdir: list[Path] = typer.Argument(
         ...,
         help="List of paths to the directory containing the Python standard library or extra packages.",
+    ),
+    exclude: str = typer.Option(
+        "",
+        help="List of files to exclude from the zip file. Defaults to no files.",
+    ),
+    stub: str = typer.Option(
+        "",
+        help="List of files that are replaced by JS implementations. Defaults to no files.",
     ),
     pycompile: bool = typer.Option(
         False, help="Whether to compile the .py files into .pyc."
@@ -23,8 +32,17 @@ def main(
     """
     Bundle Python standard libraries into a zip file.
     """
+
+    # Convert the comma / space separated strings to lists
+    excludes = [
+        item.strip() for item in re.split(r",|\s", exclude) if item.strip() != ""
+    ]
+    stubs = [item.strip() for item in re.split(r",|\s", stub) if item.strip() != ""]
+
     create_zipfile(
         libdir,
+        excludes,
+        stubs,
         output,
         pycompile=pycompile,
         filterfunc=None,