+++ pyodide-build/pyodide_build/pypabuild.py
@@ -157,7 +157,7 @@ def _build_in_isolated_env(
         installed_requires_for_build = False
         try:
             build_reqs = builder.get_requires_for_build(
-                distribution,
+                distribution,  # type: ignore[arg-type]
             )
         except BuildBackendException:
             pass
@@ -168,12 +168,16 @@ def _build_in_isolated_env(
         with common.replace_env(build_env):
             if not installed_requires_for_build:
                 build_reqs = builder.get_requires_for_build(
-                    distribution,
+                    distribution,  # type: ignore[arg-type]
                     config_settings,
                 )
                 install_reqs(env, build_reqs)
 
-            return builder.build(distribution, outdir, config_settings)
+            return builder.build(
+                distribution,  # type: ignore[arg-type]
+                outdir,
+                config_settings,
+            )
 
 
 def parse_backend_flags(backend_flags: str | list[str]) -> ConfigSettingsType: