+++ src/core/pyproxy.ts
@@ -48,6 +48,8 @@ declare var IS_GENERATOR: number;
 declare var IS_ASYNC_GENERATOR: number;
 declare var IS_SEQUENCE: number;
 declare var IS_MUTABLE_SEQUENCE: number;
+declare var IS_JSON_ADAPTOR_DICT: number;
+declare var IS_JSON_ADAPTOR_SEQUENCE: number;
 
 declare function DEREF_U32(ptr: number, offset: number): number;
 declare function Py_ENTER(): void;
@@ -122,7 +124,12 @@ Module.disable_pyproxy_allocation_tracing = function () {
 };
 Module.disable_pyproxy_allocation_tracing();
 
-type PyProxyCache = { map: Map<string, any>; refcnt: number; leaked?: boolean };
+type PyProxyCache = {
+  map: Map<string, any>;
+  json_adaptor_map: Map<string, any>;
+  refcnt: number;
+  leaked?: boolean;
+};
 type PyProxyShared = {
   ptr: number;
   cache: PyProxyCache;
@@ -162,10 +169,10 @@ type PyProxyAttrs = {
 };
 
 const pyproxyAttrsSymbol = Symbol("pyproxy.attrs");
-function pyproxy_getflags(ptrobj: number) {
+function pyproxy_getflags(ptrobj: number, is_json_adaptor: boolean) {
   Py_ENTER();
   try {
-    return _pyproxy_getflags(ptrobj);
+    return _pyproxy_getflags(ptrobj, is_json_adaptor);
   } finally {
     Py_EXIT();
   }
@@ -198,25 +205,29 @@ function pyproxy_new(
     props,
     shared,
     gcRegister,
+    jsonAdaptor,
   }: {
     flags?: number;
     cache?: PyProxyCache;
     shared?: PyProxyShared;
     props?: any;
     gcRegister?: boolean;
+    jsonAdaptor?: boolean;
   } = {},
 ): PyProxy {
   if (gcRegister === undefined) {
     // register by default
     gcRegister = true;
   }
-  const flags = flags_arg !== undefined ? flags_arg : pyproxy_getflags(ptr);
+  const flags =
+    flags_arg !== undefined ? flags_arg : pyproxy_getflags(ptr, !!jsonAdaptor);
   if (flags === -1) {
     _pythonexc2js();
   }
   const is_sequence = flags & IS_SEQUENCE;
+  const is_dict_adaptor = flags & IS_JSON_ADAPTOR_DICT;
   const cls = Module.getPyProxyClass(flags);
-  let target;
+  let target: any;
   if (flags & IS_CALLABLE) {
     // In this case we are effectively subclassing Function in order to ensure
     // that the proxy is callable. With a Content Security Protocol that doesn't
@@ -245,7 +256,7 @@ function pyproxy_new(
     // Not an alias so we have to make `shared`.
     if (!cache) {
       // In this case it's not a copy.
-      cache = { map: new Map(), refcnt: 0 };
+      cache = { map: new Map(), json_adaptor_map: new Map(), refcnt: 0 };
     }
     cache.refcnt++;
     shared = {
@@ -263,10 +274,15 @@ function pyproxy_new(
     { isBound: false, captureThis: false, boundArgs: [], roundtrip: false },
     props,
   );
-  let proxy = new Proxy(
-    target,
-    is_sequence ? PyProxySequenceHandlers : PyProxyHandlers,
-  );
+  let handlers;
+  if (is_dict_adaptor) {
+    handlers = PyProxyJsonAdaptorDictHandlers;
+  } else if (is_sequence) {
+    handlers = PyProxySequenceHandlers;
+  } else {
+    handlers = PyProxyHandlers;
+  }
+  let proxy = new Proxy(target, handlers);
   if (!isAlias && gcRegister) {
     // we need to register only once for a set of aliases. we can't register the
     // proxy directly since that isn't shared between aliases. The aliases all
@@ -311,6 +327,13 @@ function _getPtr(jsobj: any) {
 
 function _getFlags(jsobj: any): number {
   return Object.getPrototypeOf(jsobj).$$flags;
+}
+
+function isJsonAdaptor(jsobj: any): boolean {
+  return !!(
+    _getFlags(jsobj) &
+    (IS_JSON_ADAPTOR_SEQUENCE | IS_JSON_ADAPTOR_DICT)
+  );
 }
 
 function _adjustArgs(proxyobj: any, jsthis: any, jsargs: any[]): any[] {
@@ -368,6 +391,12 @@ Module.getPyProxyClass = function (flags: number) {
       );
     }
   }
+  if (flags & IS_SEQUENCE || flags & HAS_GET) {
+    Object.assign(
+      descriptors,
+      Object.getOwnPropertyDescriptors(PyAsJsonAdaptorMethods.prototype),
+    );
+  }
   // Use base constructor (just throws an error if construction is attempted).
   descriptors.constructor = Object.getOwnPropertyDescriptor(
     PyProxy.prototype,
@@ -397,11 +426,15 @@ function pyproxy_decref_cache(cache: PyProxyCache) {
     return;
   }
   cache.refcnt--;
+  if (cache.leaked) {
+    return;
+  }
   if (cache.refcnt === 0) {
-    for (let proxy of cache.map.values()) {
-      if (!cache.leaked) {
-        Module.pyproxy_destroy(proxy, pyproxy_cache_destroyed_msg, true);
-      }
+    for (const proxy of cache.map.values()) {
+      Module.pyproxy_destroy(proxy, pyproxy_cache_destroyed_msg, true);
+    }
+    for (const proxy of cache.json_adaptor_map.values()) {
+      Module.pyproxy_destroy(proxy, pyproxy_cache_destroyed_msg, true);
     }
   }
 }
@@ -472,11 +505,7 @@ Module.pyproxy_destroy = function (
 // Now a lot of boilerplate to wrap the abstract Object protocol wrappers
 // defined in pyproxy.c in JavaScript functions.
 
-Module.callPyObjectKwargs = function (
-  ptrobj: number,
-  jsargs: any[],
-  kwargs: any,
-) {
+function callPyObjectKwargs(ptrobj: number, jsargs: any[], kwargs: any) {
   // We don't do any checking for kwargs, checks are in PyProxy.callKwargs
   // which only is used when the keyword arguments come from the user.
   const num_pos_args = jsargs.length;
@@ -513,7 +542,7 @@ Module.callPyObjectKwargs = function (
     }
   }
   return result;
-};
+}
 
 /**
  * A version of callPyObjectKwargs that supports the JSPI.
@@ -527,7 +556,7 @@ Module.callPyObjectKwargs = function (
  * into suspenderGlobal (for later use by JsvPromise_syncify). Then it calls
  * _pyproxy_apply with the same arguments we gave to `promisingApply`.
  */
-async function callPyObjectKwargsSuspending(
+async function callPyObjectKwargsPromising(
   ptrobj: number,
   jsargs: any,
   kwargs: any,
@@ -586,18 +615,18 @@ async function callPyObjectKwargsSuspending(
   return result;
 }
 
-Module.callPyObjectMaybeSuspending = async function (
+Module.callPyObjectMaybePromising = async function (
   ptrobj: number,
   jsargs: any,
 ) {
   if (Module.jspiSupported) {
-    return await callPyObjectKwargsSuspending(ptrobj, jsargs, {});
+    return await callPyObjectKwargsPromising(ptrobj, jsargs, {});
   }
-  return Module.callPyObjectKwargs(ptrobj, jsargs, {});
+  return callPyObjectKwargs(ptrobj, jsargs, {});
 };
 
 Module.callPyObject = function (ptrobj: number, jsargs: any) {
-  return Module.callPyObjectKwargs(ptrobj, jsargs, {});
+  return callPyObjectKwargs(ptrobj, jsargs, {});
 };
 
 export interface PyProxy {
@@ -651,7 +680,7 @@ export class PyProxy {
   }
   /**
    * Returns `str(o)` (unless `pyproxyToStringRepr: true` was passed to
-   * :js:func:`loadPyodide` in which case it will return `repr(o)`)
+   * :js:func:`~globalThis.loadPyodide` in which case it will return `repr(o)`)
    */
   toString(): string {
     let ptrobj = _getPtr(this);
@@ -843,6 +872,23 @@ export class PyProxyWithGet extends PyProxy {
 
 export interface PyProxyWithGet extends PyGetItemMethods {}
 
+class PyAsJsonAdaptorMethods {
+  asJsonAdaptor() {
+    let { shared, props } = _getAttrs(this);
+    let flags = _getFlags(this);
+    if (flags & IS_SEQUENCE) {
+      flags |= IS_JSON_ADAPTOR_SEQUENCE;
+    } else {
+      flags |= IS_JSON_ADAPTOR_DICT;
+    }
+    return pyproxy_new(shared.ptr, {
+      shared,
+      flags,
+      props,
+    });
+  }
+}
+
 // Controlled by HAS_GET, appears for any class with __getitem__,
 // mp_subscript, or sq_item methods
 export class PyGetItemMethods {
@@ -853,11 +899,17 @@ export class PyGetItemMethods {
    * @returns The corresponding value.
    */
   get(key: any): any {
-    const ptrobj = _getPtr(this);
+    const { shared } = _getAttrs(this);
     let result;
     try {
       Py_ENTER();
-      result = __pyproxy_getitem(ptrobj, key);
+      // Cache is only used if isJsonAdaptor is true.
+      result = __pyproxy_getitem(
+        shared.ptr,
+        key,
+        shared.cache.json_adaptor_map,
+        isJsonAdaptor(this),
+      );
       Py_EXIT();
     } catch (e) {
       API.fatal_error(e);
@@ -870,6 +922,25 @@ export class PyGetItemMethods {
       }
     }
     return result;
+  }
+  /**
+   * Returns the object treated as a json adaptor.
+   *
+   * With a JsonAdaptor:
+   *  1. property access / modification / deletion is implemented with
+   *     :meth:`~object.__getitem__`, :meth:`~object.__setitem__`, and
+   *     :meth:`~object.__delitem__` respectively.
+   *  2. If an attribute is accessed and the result implements
+   *     :meth:`~object.__getitem__` then the result will also be a json
+   *     adaptor.
+   *
+   * For instance, ``JSON.stringify(proxy.asJsonAdaptor())`` acts like an
+   * inverse to Python's :py:func:`json.loads`.
+   */
+  asJsonAdaptor(): PyProxy & {} {
+    // This is just here for the docs. The actual implementation comes from
+    // PyAsJsonAdaptorMethods.
+    throw new Error("Should not happen");
   }
 }
 
@@ -983,18 +1054,26 @@ export class PyContainsMethods {
  * https://hacks.mozilla.org/2015/07/es6-in-depth-generators-continued/
  *
  */
-function* iter_helper(iterptr: number, token: {}): Generator<any> {
+function* iter_helper(
+  iterptr: number,
+  token: {},
+  proxyCache: Map<string, any>,
+  is_json_adaptor: boolean,
+): Generator<any> {
   const to_destroy = [];
   try {
     while (true) {
       Py_ENTER();
-      const item = __pyproxy_iter_next(iterptr);
+      const item = __pyproxy_iter_next(iterptr, proxyCache, is_json_adaptor);
       if (item === null) {
         break;
       }
       Py_EXIT();
       yield item;
-      if (API.isPyProxy(item)) {
+      // If it's a json adaptor, we cached the result so we don't need to
+      // destroy it (they'll get destroyed when we destroy the root).
+      // This is necessary to get JSON.stringify to work correctly.
+      if (!is_json_adaptor && API.isPyProxy(item)) {
         to_destroy.push(item);
       }
     }
@@ -1045,12 +1124,12 @@ export class PyIterableMethods {
    * This will be used implicitly by ``for(let x of proxy){}``.
    */
   [Symbol.iterator](): Iterator<any, any, any> {
-    let ptrobj = _getPtr(this);
+    const { shared } = _getAttrs(this);
     let token = {};
     let iterptr;
     try {
       Py_ENTER();
-      iterptr = _PyObject_GetIter(ptrobj);
+      iterptr = _PyObject_GetIter(shared.ptr);
       Py_EXIT();
     } catch (e) {
       API.fatal_error(e);
@@ -1059,7 +1138,13 @@ export class PyIterableMethods {
       _pythonexc2js();
     }
 
-    let result = iter_helper(iterptr, token);
+    // Cache is only used if isJsonAdaptor is true.
+    let result = iter_helper(
+      iterptr,
+      token,
+      shared.cache.json_adaptor_map,
+      isJsonAdaptor(this),
+    );
     Module.finalizationRegistry.register(result, [iterptr, undefined], token);
     return result;
   }
@@ -1189,7 +1274,7 @@ export class PyIteratorMethods {
    *
    * This will be used implicitly by ``for(let x of proxy){}``.
    *
-   * @param any The value to send to the generator. The value will be assigned
+   * @param arg The value to send to the generator. The value will be assigned
    * as a result of a yield expression.
    * @returns An Object with two properties: ``done`` and ``value``. When the
    * generator yields ``some_value``, ``next`` returns ``{done : false, value :
@@ -1234,7 +1319,7 @@ export class PyGeneratorMethods {
    *
    * See the documentation for :js:meth:`Generator.throw`.
    *
-   * @param exception Error The error to throw into the generator. Must be an
+   * @param exc Error The error to throw into the generator. Must be an
    * instanceof ``Error``.
    * @returns An Object with two properties: ``done`` and ``value``. When the
    * generator yields ``some_value``, ``return`` returns ``{done : false, value
@@ -1266,7 +1351,7 @@ export class PyGeneratorMethods {
    * :py:exc:`GeneratorExit` or :py:exc:`StopIteration`, that error is propagated. See
    * the documentation for :js:meth:`Generator.return`.
    *
-   * @param any The value to return from the generator.
+   * @param v The value to return from the generator.
    * @returns An Object with two properties: ``done`` and ``value``. When the
    * generator yields ``some_value``, ``return`` returns ``{done : false, value
    * : some_value}``. When the generator raises a
@@ -1316,7 +1401,7 @@ export class PyAsyncIteratorMethods {
    *
    * This will be used implicitly by ``for(let x of proxy){}``.
    *
-   * @param any The value to send to a generator. The value will be assigned as
+   * @param arg The value to send to a generator. The value will be assigned as
    * a result of a yield expression.
    * @returns An Object with two properties: ``done`` and ``value``. When the
    * iterator yields ``some_value``, ``next`` returns ``{done : false, value :
@@ -1374,7 +1459,7 @@ export class PyAsyncGeneratorMethods {
    *
    * See the documentation for :js:meth:`AsyncGenerator.throw`.
    *
-   * @param exception Error The error to throw into the generator. Must be an
+   * @param exc Error The error to throw into the generator. Must be an
    * instanceof ``Error``.
    * @returns An Object with two properties: ``done`` and ``value``. When the
    * generator yields ``some_value``, ``return`` returns ``{done : false, value
@@ -1421,7 +1506,7 @@ export class PyAsyncGeneratorMethods {
    * :py:exc:`GeneratorExit` or :py:exc:`StopAsyncIteration`, that error is
    * propagated. See the documentation for :js:meth:`AsyncGenerator.throw`
    *
-   * @param any The value to return from the generator.
+   * @param v The value to return from the generator.
    * @returns An Object with two properties: ``done`` and ``value``. When the
    * generator yields ``some_value``, ``return`` returns ``{done : false, value
    * : some_value}``. When the generator raises a :py:exc:`StopAsyncIteration`
@@ -1571,7 +1656,7 @@ export class PySequenceMethods {
    * See :js:meth:`Array.filter`. Creates a shallow copy of a portion of a given
    * ``Sequence``, filtered down to just the elements from the given array that pass
    * the test implemented by the provided function.
-   * @param callbackfn A function to execute for each element in the array. It
+   * @param predicate A function to execute for each element in the array. It
    * should return a truthy value to keep the element in the resulting array,
    * and a falsy value otherwise.
    * @param thisArg A value to use as ``this`` when executing ``predicate``.
@@ -1585,7 +1670,7 @@ export class PySequenceMethods {
   /**
    * See :js:meth:`Array.some`. Tests whether at least one element in the
    * ``Sequence`` passes the test implemented by the provided function.
-   * @param callbackfn A function to execute for each element in the
+   * @param predicate A function to execute for each element in the
    * ``Sequence``. It should return a truthy value to indicate the element
    * passes the test, and a falsy value otherwise.
    * @param thisArg A value to use as ``this`` when executing ``predicate``.
@@ -1599,7 +1684,7 @@ export class PySequenceMethods {
   /**
    * See :js:meth:`Array.every`. Tests whether every element in the ``Sequence``
    * passes the test implemented by the provided function.
-   * @param callbackfn A function to execute for each element in the
+   * @param predicate A function to execute for each element in the
    * ``Sequence``. It should return a truthy value to indicate the element
    * passes the test, and a falsy value otherwise.
    * @param thisArg A value to use as ``this`` when executing ``predicate``.
@@ -1617,7 +1702,6 @@ export class PySequenceMethods {
    * running the reducer across all elements of the Sequence is a single value.
    * @param callbackfn A function to execute for each element in the ``Sequence``. Its
    * return value is discarded.
-   * @param thisArg A value to use as ``this`` when executing ``callbackfn``.
    */
   reduce(
     callbackfn: (
@@ -1638,7 +1722,6 @@ export class PySequenceMethods {
    * single value.
    * @param callbackfn A function to execute for each element in the Sequence.
    * Its return value is discarded.
-   * @param thisArg A value to use as ``this`` when executing ``callbackFn``.
    */
   reduceRight(
     callbackfn: (
@@ -1740,6 +1823,30 @@ export class PySequenceMethods {
     thisArg?: any,
   ): number {
     return Array.prototype.findIndex.call(this, predicate, thisArg);
+  }
+
+  toJSON(this: any) {
+    return Array.from(this);
+  }
+
+  /**
+   * Returns the object treated as a json adaptor.
+   *
+   * With a JsonAdaptor:
+   *  1. property access / modification / deletion is implemented with
+   *     :meth:`~object.__getitem__`, :meth:`~object.__setitem__`, and
+   *     :meth:`~object.__delitem__` respectively.
+   *  2. If an attribute is accessed and the result implements
+   *     :meth:`~object.__getitem__` then the result will also be a json
+   *     adaptor.
+   *
+   * For instance, ``JSON.stringify(proxy.asJsonAdaptor())`` acts like an
+   * inverse to Python's :py:func:`json.loads`.
+   */
+  asJsonAdaptor(): PyProxy & {} {
+    // This is just here for the docs. The actual implementation comes from
+    // PyAsJsonAdaptorMethods.
+    throw new Error("Should not happen");
   }
 }
 
@@ -2220,6 +2327,99 @@ const PyProxySequenceHandlers = {
   },
 };
 
+const PyProxyJsonAdaptorDictHandlers = {
+  isExtensible(): boolean {
+    return true;
+  },
+  has(jsobj: PyProxy, jskey: any): boolean {
+    if (PyContainsMethods.prototype.has.call(jsobj, jskey)) {
+      return true;
+    }
+    if (typeof jskey === "string" && /^[0-9]*$/.test(jskey)) {
+      jskey = Number(jskey);
+    }
+    if (PyContainsMethods.prototype.has.call(jsobj, jskey)) {
+      return true;
+    }
+    return false;
+  },
+  get(jsobj: PyProxy, jskey: any): any {
+    let result;
+    if (
+      ["copy", "constructor", "$$flags", "toString", "destroy"].includes(
+        jskey,
+      ) ||
+      typeof jskey === "symbol"
+    ) {
+      // @ts-ignore
+      return Reflect.get(...arguments);
+    }
+    if (typeof jskey === "string") {
+      // TODO: consider adding an attribute cache for asJsonAdaptor
+      result = PyGetItemMethods.prototype.get.call(jsobj, jskey);
+    }
+    if (result) {
+      return result;
+    }
+    if (typeof jskey === "string" && /^[0-9]*$/.test(jskey)) {
+      jskey = Number(jskey);
+      result = PyGetItemMethods.prototype.get.call(jsobj, jskey);
+    }
+    if (result) {
+      return result;
+    }
+    // @ts-ignore
+    return Reflect.get(...arguments);
+  },
+  set(jsobj: PyProxy, jskey: any, jsval: any): boolean {
+    if (typeof jskey === "string") {
+      try {
+        PySetItemMethods.prototype.set.call(jsobj, Number(jskey), jsval);
+        return true;
+      } catch (e) {
+        if (isPythonError(e)) {
+          return false;
+        }
+        throw e;
+      }
+    }
+    return false;
+  },
+  deleteProperty(jsobj: PyProxy, jskey: any): boolean {
+    if (typeof jskey === "string" && /^[0-9]*$/.test(jskey)) {
+      try {
+        PySetItemMethods.prototype.delete.call(jsobj, Number(jskey));
+        return true;
+      } catch (e) {
+        if (isPythonError(e)) {
+          return false;
+        }
+        throw e;
+      }
+    }
+    return false;
+  },
+  getOwnPropertyDescriptor(jsobj: PyProxy, prop: any) {
+    if (!PyProxyJsonAdaptorDictHandlers.has(jsobj, prop)) {
+      return undefined;
+    }
+    const value = PyProxyJsonAdaptorDictHandlers.get(jsobj, prop);
+    return {
+      configurable: true,
+      enumerable: true,
+      value,
+      writable: true,
+    };
+  },
+  ownKeys(jsobj: PyProxy): (string | symbol)[] {
+    let dict_keys_view: Iterable<string> & PyProxy;
+    dict_keys_view = PyProxyHandlers.get(jsobj, "keys")();
+    const res = Array.from(dict_keys_view);
+    dict_keys_view.destroy();
+    return res;
+  },
+};
+
 /**
  * A :js:class:`~pyodide.ffi.PyProxy` whose proxied Python object is :ref:`awaitable
  * <asyncio-awaitables>` (i.e., has an :meth:`~object.__await__` method).
@@ -2380,6 +2580,55 @@ export class PyCallableMethods {
     jsargs = _adjustArgs(this, thisArg, jsargs);
     return Module.callPyObject(_getPtr(this), jsargs);
   }
+
+  /**
+   * Call the Python function. The first parameter controls various parameters
+   * that change the way the call is performed.
+   *
+   * @param options
+   * @param options.kwargs If true, the last argument is treated as a collection
+   *                       of keyword arguments.
+   * @param options.promising If true, the call is made with stack switching
+   *                          enabled. Not needed if the callee is an async
+   *                          Python function.
+   * @param options.relaxed If true, extra arguments are ignored instead of
+   *                        raising a :py:exc:`TypeError`.
+   * @param jsargs Arguments to the Python function.
+   * @returns
+   */
+  callWithOptions(
+    {
+      relaxed,
+      kwargs,
+      promising,
+    }: { relaxed?: boolean; kwargs?: boolean; promising?: boolean },
+    ...jsargs: any
+  ) {
+    let kwarg = {};
+    if (kwargs) {
+      if (jsargs.length === 0) {
+        throw new TypeError(
+          "callWithOptions with 'kwargs: true' requires at least one argument (the key word argument object)",
+        );
+      }
+      kwarg = jsargs.pop();
+      if (
+        kwarg.constructor !== undefined &&
+        kwarg.constructor.name !== "Object"
+      ) {
+        throw new TypeError("kwargs argument is not an object");
+      }
+    }
+    const target = relaxed ? API.pyodide_code.relaxed_call : this;
+    if (relaxed) {
+      jsargs.unshift(this);
+    }
+    const callFunc = promising
+      ? callPyObjectKwargsPromising
+      : callPyObjectKwargs;
+    return callFunc(_getPtr(target), jsargs, kwarg);
+  }
+
   /**
    * Call the function with keyword arguments. The last argument must be an
    * object with the keyword arguments.
@@ -2397,7 +2646,7 @@ export class PyCallableMethods {
     ) {
       throw new TypeError("kwargs argument is not an object");
     }
-    return Module.callPyObjectKwargs(_getPtr(this), jsargs, kwargs);
+    return callPyObjectKwargs(_getPtr(this), jsargs, kwargs);
   }
 
   /**
@@ -2421,8 +2670,8 @@ export class PyCallableMethods {
    * will be ignored. This matches the behavior of JavaScript functions more
    * accurately.
    *
-   * Missing arguments are **NOT** filled with `None`. If too few arguments are
-   * passed, this will still raise a TypeError. Also, if the same argument is
+   * Missing arguments are **NOT** filled with ``None``. If too few arguments are
+   * passed, this will still raise a :py:exc:`TypeError`. Also, if the same argument is
    * passed as both a keyword argument and a positional argument, it will raise
    * an error.
    *
@@ -2433,11 +2682,11 @@ export class PyCallableMethods {
   }
 
   /**
-   * Call the function with stack switching enabled. Functions called this way
-   * can use
-   * :py:meth:`PyodideFuture.syncify() <pyodide.webloop.PyodideFuture.syncify>`
-   * to block until a :py:class:`~asyncio.Future` or :js:class:`Promise` is
-   * resolved. Only works in runtimes with JS Promise integration.
+   * Call the function with stack switching enabled. The last argument must be
+   * an object with the keyword arguments. Functions called this way can use
+   * :py:meth:`~pyodide.ffi.run_sync` to block until an
+   * :py:class:`~collections.abc.Awaitable` is resolved. Only works in runtimes
+   * with JS Promise integration.
    *
    * .. admonition:: Experimental
    *    :class: warning
@@ -2446,16 +2695,16 @@ export class PyCallableMethods {
    *
    * @experimental
    */
-  callSyncifying(...jsargs: any) {
-    return callPyObjectKwargsSuspending(_getPtr(this), jsargs, {});
+  callPromising(...jsargs: any) {
+    return callPyObjectKwargsPromising(_getPtr(this), jsargs, {});
   }
 
   /**
    * Call the function with stack switching enabled. The last argument must be
    * an object with the keyword arguments. Functions called this way can use
-   * :py:meth:`PyodideFuture.syncify() <pyodide.webloop.PyodideFuture.syncify>`
-   * to block until a :py:class:`~asyncio.Future` or :js:class:`Promise` is
-   * resolved. Only works in runtimes with JS Promise integration.
+   * :py:meth:`~pyodide.ffi.run_sync` to block until an
+   * :py:class:`~collections.abc.Awaitable` is resolved. Only works in runtimes
+   * with JS Promise integration.
    *
    * .. admonition:: Experimental
    *    :class: warning
@@ -2464,7 +2713,7 @@ export class PyCallableMethods {
    *
    * @experimental
    */
-  callSyncifyingKwargs(...jsargs: any) {
+  callPromisingKwargs(...jsargs: any) {
     if (jsargs.length === 0) {
       throw new TypeError(
         "callKwargs requires at least one argument (the key word argument object)",
@@ -2477,7 +2726,7 @@ export class PyCallableMethods {
     ) {
       throw new TypeError("kwargs argument is not an object");
     }
-    return callPyObjectKwargsSuspending(_getPtr(this), jsargs, kwargs);
+    return callPyObjectKwargsPromising(_getPtr(this), jsargs, kwargs);
   }
 
   /**
@@ -2897,8 +3146,13 @@ export class PyBufferView {
    */
   f_contiguous: boolean;
 
+  /**
+   * @private
+   */
   _released: boolean;
-
+  /**
+   * @private
+   */
   _view_ptr: number;
 
   /** @private */